Modify type of AtomKey from uint32_t to uint16_t

+++ include/utils/AtomKeySystem.hpp
@@ -7,7 +7,7 @@
 
 #include "GlobalEnumClass.hpp"
 
-using AtomKey = uint32_t;
+using AtomKey = uint16_t;
 
 class AtomKeySystem
 {
